The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Dataset

daily summary of the Coronavirus (COVID-19) cases by state/province.

coronavirus

Format

A tbl object

Source

Johns Hopkins University Center for Systems Science and Engineering (JHU CCSE) Coronavirus website

Details

The dataset contains the daily summary of Coronavirus cases (confirmed, death, and recovered), by state/province

Examples

data(coronavirus)

require(dplyr)

#> Loading required package: dplyr

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# Get top confirmed cases by state

coronavirus %>%
 filter(type == "confirmed") %>%
 group_by(Country.Region) %>%
 summarise(total = sum(cases)) %>%
 arrange(-total) %>%
 head(20)

#> # A tibble: 20 x 2
#>    Country.Region total
#>    <chr>          <int>
#>  1 Mainland China 80699
#>  2 Italy           7375
#>  3 South Korea     7314
#>  4 Iran            6566
#>  5 France          1126
#>  6 Germany         1040
#>  7 Others           696
#>  8 Spain            673
#>  9 US               538
#> 10 Japan            502
#> 11 Switzerland      337
#> 12 UK               273
#> 13 Netherlands      265
#> 14 Sweden           203
#> 15 Belgium          200
#> 16 Norway           176
#> 17 Singapore        150
#> 18 Hong Kong        114
#> 19 Austria          104
#> 20 Malaysia          99

# Get the number of recovered cases in Mainland China by province
 coronavirus %>%
    filter(type == "recovered", Country.Region == "Mainland China") %>%
    group_by(Province.State) %>%
    summarise(total = sum(cases)) %>%
    arrange(-total)

#> # A tibble: 31 x 2
#>    Province.State total
#>    <chr>          <int>
#>  1 Hubei          45235
#>  2 Guangdong       1256
#>  3 Henan           1247
#>  4 Zhejiang        1161
#>  5 Anhui            984
#>  6 Hunan            968
#>  7 Jiangxi          919
#>  8 Shandong         642
#>  9 Jiangsu          612
#> 10 Chongqing        527
#> # … with 21 more rows