The COVID-19 Worldwide Vaccine Dataset

Daily summary of the COVID-19 vaccination by country/province.

covid19_vaccine

Format

A data frame with 8 variables.

country_region

Country or region name

date

Data collection date in YYYY-MM-DD format

doses_admin

Cumulative number of doses administered. When a vaccine requires multiple doses, each one is counted independently

people_partially_vaccinated

Cumulative number of people who received at least one vaccine dose. When the person receives a prescribed second dose, it is not counted twice

people_fully_vaccinated

Cumulative number of people who received all prescribed doses necessary to be considered fully vaccinated

report_date_string

Data report date in YYYY-MM-DD format

uid

Country code

province_state

Province or state if applicable

iso2

Officially assigned country code identifiers with two-letter

iso3

Officially assigned country code identifiers with three-letter

code3

UN country code

fips

Federal Information Processing Standards code that uniquely identifies counties within the USA

lat

Latitude

long

Longitude

combined_key

Country and province (if applicable)

population

Country or province population

continent_name

Continent name

continent_code

Continent code

Source

- Vaccine data - Johns Hopkins University Centers for Civic Impact (JHU CCSE) COVID-19 repository.

- Country code (uid, iso2, iso3, etc.) are sourced from this repository, see section 4 for full data resources.

- Continent code mapping is sourced from DATA HUB

Details

The dataset provides the daily cumulative number of people who received vaccine (or at least one vaccine dose) by country and province (when applicable)

Examples

data(covid19_vaccine)

head(covid19_vaccine)

#>   country_region       date doses_admin people_partially_vaccinated
#> 1    Afghanistan 2021-02-22           0                           0
#> 2    Afghanistan 2021-02-23           0                           0
#> 3    Afghanistan 2021-02-24           0                           0
#> 4    Afghanistan 2021-02-25           0                           0
#> 5    Afghanistan 2021-02-26           0                           0
#> 6    Afghanistan 2021-02-27           0                           0
#>   people_fully_vaccinated report_date_string uid province_state iso2 iso3 code3
#> 1                       0         2021-02-22   4           <NA>   AF  AFG     4
#> 2                       0         2021-02-23   4           <NA>   AF  AFG     4
#> 3                       0         2021-02-24   4           <NA>   AF  AFG     4
#> 4                       0         2021-02-25   4           <NA>   AF  AFG     4
#> 5                       0         2021-02-26   4           <NA>   AF  AFG     4
#> 6                       0         2021-02-27   4           <NA>   AF  AFG     4
#>   fips      lat     long combined_key population continent_name continent_code
#> 1 <NA> 33.93911 67.70995  Afghanistan   38928341           Asia             AS
#> 2 <NA> 33.93911 67.70995  Afghanistan   38928341           Asia             AS
#> 3 <NA> 33.93911 67.70995  Afghanistan   38928341           Asia             AS
#> 4 <NA> 33.93911 67.70995  Afghanistan   38928341           Asia             AS
#> 5 <NA> 33.93911 67.70995  Afghanistan   38928341           Asia             AS
#> 6 <NA> 33.93911 67.70995  Afghanistan   38928341           Asia             AS