The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Dataset

Daily summary of the Coronavirus (COVID-19) cases by state/province.

coronavirus

Format

A data frame with 7 variables.

date

Date in YYYY-MM-DD format.

province

Name of province/state, for countries where data is provided split across multiple provinces/states.

country

Name of country/region.

lat

Latitude of center of geographic region, defined as either country or, if available, province.

long

Longitude of center of geographic region, defined as either country or, if available, province.

type

An indicator for the type of cases (confirmed, death, recovered).

cases

Number of cases on given date.

Source

Johns Hopkins University Center for Systems Science and Engineering (JHU CCSE) Coronavirus website.

Details

The dataset contains the daily summary of Coronavirus cases (confirmed, death, and recovered), by state/province.

Examples

data(coronavirus)

require(dplyr)

#> Loading required package: dplyr

#> Warning: package ‘dplyr’ was built under R version 3.6.2

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# Get top confirmed cases by state
coronavirus %>%
  filter(type == "confirmed") %>%
  group_by(country) %>%
  summarise(total = sum(cases)) %>%
  arrange(-total) %>%
  head(20)

#> # A tibble: 20 x 2
#>    country           total
#>    <chr>             <dbl>
#>  1 US             33166418
#>  2 India          27157795
#>  3 Brazil         16194209
#>  4 France          5670486
#>  5 Turkey          5203385
#>  6 Russia          4960174
#>  7 United Kingdom  4483177
#>  8 Italy           4197892
#>  9 Germany         3662568
#> 10 Spain           3652879
#> 11 Argentina       3586736
#> 12 Colombia        3270614
#> 13 Poland          2867187
#> 14 Iran            2855396
#> 15 Mexico          2399790
#> 16 Ukraine         2244084
#> 17 Peru            1932255
#> 18 Indonesia       1786187
#> 19 Czechia         1658778
#> 20 Netherlands     1658587

# Get the number of recovered cases in China by province
coronavirus %>%
  filter(type == "recovered", country == "China") %>%
  group_by(province) %>%
  summarise(total = sum(cases)) %>%
  arrange(-total)

#> # A tibble: 34 x 2
#>    province     total
#>    <chr>        <dbl>
#>  1 Hubei        63645
#>  2 Hong Kong    11560
#>  3 Guangdong     2364
#>  4 Shanghai      2004
#>  5 Heilongjiang  1597
#>  6 Zhejiang      1324
#>  7 Hebei         1310
#>  8 Henan         1290
#>  9 Beijing       1046
#> 10 Hunan         1039
#> # … with 24 more rows