The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Dataset

Daily summary of the Coronavirus (COVID-19) cases by state/province.

coronavirus

Format

A data frame with 7 variables.

date

Date in YYYY-MM-DD format.

province

Name of province/state, for countries where data is provided split across multiple provinces/states.

country

Name of country/region.

lat

Latitude of center of geographic region, defined as either country or, if available, province.

long

Longitude of center of geographic region, defined as either country or, if available, province.

type

An indicator for the type of cases (confirmed, death, recovered).

cases

Number of cases on given date.

Source

Johns Hopkins University Center for Systems Science and Engineering (JHU CCSE) Coronavirus website.

Details

The dataset contains the daily summary of Coronavirus cases (confirmed, death, and recovered), by state/province.

Examples

data(coronavirus)

require(dplyr)

#> Loading required package: dplyr

#> Warning: package ‘dplyr’ was built under R version 3.6.2

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# Get top confirmed cases by state
coronavirus %>%
  filter(type == "confirmed") %>%
  group_by(country) %>%
  summarise(total = sum(cases)) %>%
  arrange(-total) %>%
  head(20)

#> `summarise()` ungrouping output (override with `.groups` argument)

#> # A tibble: 20 x 2
#>    country           total
#>    <chr>             <int>
#>  1 US             21574043
#>  2 India          10413417
#>  3 Brazil          7961673
#>  4 Russia          3297833
#>  5 United Kingdom  2898052
#>  6 France          2763370
#>  7 Turkey          2296102
#>  8 Italy           2220361
#>  9 Spain           2024904
#> 10 Germany         1886561
#> 11 Colombia        1737347
#> 12 Argentina       1690006
#> 13 Mexico          1493569
#> 14 Poland          1356882
#> 15 Iran            1268263
#> 16 South Africa    1170590
#> 17 Ukraine         1133802
#> 18 Peru            1022018
#> 19 Netherlands      862805
#> 20 Indonesia        797723

# Get the number of recovered cases in China by province
coronavirus %>%
  filter(type == "recovered", country == "China") %>%
  group_by(province) %>%
  summarise(total = sum(cases)) %>%
  arrange(-total)

#> `summarise()` ungrouping output (override with `.groups` argument)

#> # A tibble: 33 x 2
#>    province     total
#>    <chr>        <int>
#>  1 Hubei        63637
#>  2 Hong Kong     8258
#>  3 Guangdong     2020
#>  4 Shanghai      1442
#>  5 Zhejiang      1296
#>  6 Henan         1273
#>  7 Hunan         1016
#>  8 Anhui          986
#>  9 Xinjiang       977
#> 10 Heilongjiang   949
#> # … with 23 more rows