The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Dataset

Daily summary of the Coronavirus (COVID-19) cases by state/province.

coronavirus

Format

A data frame with 7 variables.

date

Date in YYYY-MM-DD format.

province

Name of province/state, for countries where data is provided split across multiple provinces/states.

country

Name of country/region.

lat

Latitude of center of geographic region, defined as either country or, if available, province.

long

Longitude of center of geographic region, defined as either country or, if available, province.

type

An indicator for the type of cases (confirmed, death, recovered).

cases

Number of cases on given date.

uid

Country code

iso2

Officially assigned country code identifiers with two-letter

iso3

Officially assigned country code identifiers with three-letter

code3

UN country code

combined_key

Country and province (if applicable)

population

Country or province population

continent_name

Continent name

continent_code

Continent code

Source

Johns Hopkins University Center for Systems Science and Engineering (JHU CCSE) Coronavirus website.

Details

The dataset contains the daily summary of Coronavirus cases (confirmed, death, and recovered), by state/province.

Examples

data(coronavirus)

require(dplyr)

#> Loading required package: dplyr

#> Warning: package ‘dplyr’ was built under R version 3.6.2

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# Get top confirmed cases by state
coronavirus %>%
  filter(type == "confirmed") %>%
  group_by(country) %>%
  summarise(total = sum(cases)) %>%
  arrange(-total) %>%
  head(20)

#> # A tibble: 20 × 2
#>    country           total
#>    <chr>             <int>
#>  1 US             44339747
#>  2 India          33971607
#>  3 Brazil         21575820
#>  4 United Kingdom  8192589
#>  5 Russia          7658923
#>  6 Turkey          7444522
#>  7 France          7156066
#>  8 Iran            5702890
#>  9 Argentina       5265859
#> 10 Spain           4973619
#> 11 Colombia        4972236
#> 12 Italy           4700316
#> 13 Germany         4318437
#> 14 Indonesia       4227932
#> 15 Mexico          3723235
#> 16 Poland          2922401
#> 17 South Africa    2912145
#> 18 Philippines     2666562
#> 19 Ukraine         2657699
#> 20 Malaysia        2339594

# Get the number of recovered cases in China by province
coronavirus %>%
  filter(type == "recovered", country == "China") %>%
  group_by(province) %>%
  summarise(total = sum(cases)) %>%
  arrange(-total)

#> # A tibble: 0 × 2
#> # … with 2 variables: province <chr>, total <int>