The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Dataset

Daily summary of the Coronavirus (COVID-19) cases by state/province.

coronavirus

Format

A data frame with 7 variables.

date

Date in YYYY-MM-DD format.

province

Name of province/state, for countries where data is provided split across multiple provinces/states.

country

Name of country/region.

lat

Latitude of center of geographic region, defined as either country or, if available, province.

long

Longitude of center of geographic region, defined as either country or, if available, province.

type

An indicator for the type of cases (confirmed, death, recovered).

cases

Number of cases on given date.

uid

Country code

iso2

Officially assigned country code identifiers with two-letter

iso3

Officially assigned country code identifiers with three-letter

code3

UN country code

combined_key

Country and province (if applicable)

population

Country or province population

continent_name

Continent name

continent_code

Continent code

Source

Johns Hopkins University Center for Systems Science and Engineering (JHU CCSE) Coronavirus website.

Details

The dataset contains the daily summary of Coronavirus cases (confirmed, death, and recovered), by state/province.

Examples

data(coronavirus)

require(dplyr)
#> Loading required package: dplyr
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# Get top confirmed cases by state
coronavirus %>%
  filter(type == "confirmed") %>%
  group_by(country) %>%
  summarise(total = sum(cases)) %>%
  arrange(-total) %>%
  head(20)
#> # A tibble: 20 × 2
#>    country            total
#>    <chr>              <dbl>
#>  1 US             102870369
#>  2 India           44685499
#>  3 France          39770404
#>  4 Germany         37928944
#>  5 Brazil          36953492
#>  6 Japan           32980916
#>  7 Korea, South    30369744
#>  8 Italy           25519067
#>  9 United Kingdom  24549097
#> 10 Russia          21773482
#> 11 Turkey          17042722
#> 12 Spain           13748918
#> 13 Vietnam         11526704
#> 14 Australia       11330342
#> 15 Argentina       10042136
#> 16 Taiwan*          9821755
#> 17 Netherlands      8699058
#> 18 Iran             7565678
#> 19 Mexico           7400848
#> 20 Indonesia        6732968

# Get the number of recovered cases in China by province
coronavirus %>%
  filter(type == "recovered", country == "China") %>%
  group_by(province) %>%
  summarise(total = sum(cases)) %>%
  arrange(-total)
#> # A tibble: 0 × 2
#> # … with 2 variables: province <chr>, total <dbl>