The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Dataset

daily summary of the Coronavirus (COVID-19) cases by state/province.

coronavirus

Format

A data.frame object

Source

Johns Hopkins University Center for Systems Science and Engineering (JHU CCSE) Coronavirus website

Details

The dataset contains the daily summary of Coronavirus cases (confirmed, death, and recovered), by state/province

Examples

data(coronavirus)

require(dplyr)

#> Loading required package: dplyr

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# Get top confirmed cases by state

coronavirus %>%
 filter(type == "confirmed") %>%
 group_by(Country.Region) %>%
 summarise(total = sum(cases)) %>%
 arrange(-total) %>%
 head(20)

#> # A tibble: 20 x 2
#>    Country.Region  total
#>    <chr>           <int>
#>  1 US             308850
#>  2 Spain          126168
#>  3 Italy          124632
#>  4 Germany         96092
#>  5 France          90848
#>  6 China           82543
#>  7 Iran            55743
#>  8 United Kingdom  42477
#>  9 Turkey          23934
#> 10 Switzerland     20505
#> 11 Belgium         18431
#> 12 Netherlands     16727
#> 13 Canada          12978
#> 14 Austria         11781
#> 15 Portugal        10524
#> 16 Brazil          10360
#> 17 Korea, South    10156
#> 18 Israel           7851
#> 19 Sweden           6443
#> 20 Australia        5550

# Get the number of recovered cases in Mainland China by province
 coronavirus %>%
    filter(type == "recovered", Country.Region == "Mainland China") %>%
    group_by(Province.State) %>%
    summarise(total = sum(cases)) %>%
    arrange(-total)

#> # A tibble: 0 x 2
#> # … with 2 variables: Province.State <chr>, total <int>