The COVID-19 Worldwide Vaccine Dataset

Daily summary of the COVID-19 vaccination by country/province.

covid19_vaccine

Format

A data frame with 8 variables.

country_region

Country or region name

date

Data collection date in YYYY-MM-DD format

doses_admin

Cumulative number of doses administered. When a vaccine requires multiple doses, each one is counted independently

people_partially_vaccinated

Cumulative number of people who received at least one vaccine dose. When the person receives a prescribed second dose, it is not counted twice

people_fully_vaccinated

Cumulative number of people who received all prescribed doses necessary to be considered fully vaccinated

report_date_string

Data report date in YYYY-MM-DD format

uid

Country code

province_state

Province or state if applicable

iso2

Officially assigned country code identifiers with two-letter

iso3

Officially assigned country code identifiers with three-letter

code3

UN country code

fips

Federal Information Processing Standards code that uniquely identifies counties within the USA

lat

Latitude

long

Longitude

combined_key

Country and province (if applicable)

population

Country or province population

continent_name

Continent name

continent_code

Continent code

Source

- Vaccine data - Johns Hopkins University Centers for Civic Impact (JHU CCSE) COVID-19 repository.

- Country code (uid, iso2, iso3, etc.) are sourced from this repository, see section 4 for full data resources.

- Continent code mapping is sourced from DATA HUB

Details

The dataset provides the daily cumulative number of people who received vaccine (or at least one vaccine dose) by country and province (when applicable)

Examples

data(covid19_vaccine)

head(covid19_vaccine)
#>   country_region       date doses_admin people_partially_vaccinated
#> 1         Canada 2020-12-14           5                           0
#> 2          World 2020-12-14           5                           0
#> 3         Canada 2020-12-15         723                           0
#> 4          China 2020-12-15     1500000                           0
#> 5         Russia 2020-12-15       28500                       28500
#> 6          World 2020-12-15     1529223                       28500
#>   people_fully_vaccinated report_date_string uid province_state iso2 iso3 code3
#> 1                       0         2020-12-14 124           <NA>   CA  CAN   124
#> 2                       0         2020-12-14  NA           <NA> <NA> <NA>    NA
#> 3                       0         2020-12-15 124           <NA>   CA  CAN   124
#> 4                       0         2020-12-15 156           <NA>   CN  CHN   156
#> 5                       0         2020-12-15 643           <NA>   RU  RUS   643
#> 6                       0         2020-12-15  NA           <NA> <NA> <NA>    NA
#>   fips      lat     long combined_key population continent_name continent_code
#> 1 <NA> 60.00000 -95.0000       Canada   37855702  North America             NA
#> 2 <NA>       NA       NA         <NA>         NA           <NA>           <NA>
#> 3 <NA> 60.00000 -95.0000       Canada   37855702  North America             NA
#> 4 <NA> 35.86170 104.1954        China 1404676330           Asia             AS
#> 5 <NA> 61.52401 105.3188       Russia  145934460         Europe             EU
#> 6 <NA>       NA       NA         <NA>         NA           <NA>           <NA>