The COVID-19 Worldwide Vaccine Dataset

Daily summary of the COVID-19 vaccination by country/province.

covid19_vaccine

Format

A data frame with 8 variables.

date

Data collection date in YYYY-MM-DD format

country_region

Country or region name

continent_name

Continent name

continent_code

Continent code

combined_key

Country and province (if applicable)

doses_admin

Cumulative number of doses administered. When a vaccine requires multiple doses, each one is counted independently

people_at_least_one_dose

Cumulative number of people who received at least one vaccine dose. When the person receives a prescribed second dose, it is not counted twice

population

Country or province population

uid

Country code

iso2

Officially assigned country code identifiers with two-letter

iso3

Officially assigned country code identifiers with three-letter

code3

UN country code

fips

Federal Information Processing Standards code that uniquely identifies counties within the USA

lat

Latitude

long

Longitude

Source

- Vaccine data - Johns Hopkins University Centers for Civic Impact (JHU CCSE) COVID-19 repository.

- Country code (uid, iso2, iso3, etc.) are sourced from this repository, see section 4 for full data resources.

- Continent code mapping is sourced from DATA HUB

Details

The dataset provides the daily cumulative number of people who received vaccine (or at least one vaccine dose) by country and province (when applicable)

Examples

data(covid19_vaccine)

head(covid19_vaccine)
#>         date country_region continent_name continent_code combined_key
#> 1 2020-12-29        Austria         Europe             EU      Austria
#> 2 2020-12-29        Bahrain           Asia             AS      Bahrain
#> 3 2020-12-29        Belarus         Europe             EU      Belarus
#> 4 2020-12-29        Belgium         Europe             EU      Belgium
#> 5 2020-12-29         Canada  North America             NA       Canada
#> 6 2020-12-29          Chile  South America             SA        Chile
#>   doses_admin people_at_least_one_dose population uid iso2 iso3 code3 fips
#> 1        2123                     2123    9006400  40   AT  AUT    40 <NA>
#> 2       55014                    55014    1701583  48   BH  BHR    48 <NA>
#> 3           0                        0    9449321 112   BY  BLR   112 <NA>
#> 4         340                      340   11589616  56   BE  BEL    56 <NA>
#> 5       59079                    59078   37855702 124   CA  CAN   124 <NA>
#> 6          NA                       NA   19116209 152   CL  CHL   152 <NA>
#>        lat       long
#> 1  47.5162  14.550100
#> 2  26.0275  50.550000
#> 3  53.7098  27.953400
#> 4  50.8333   4.469936
#> 5  60.0000 -95.000000
#> 6 -35.6751 -71.543000